fix(build): make windows cross-build compile
The release `binary` job's `make windows` target failed to compile
because `ui/dev.go` and `cmd/connection_browser*.go` used POSIX-only
syscalls (`SysProcAttr.Setpgid`, `syscall.Getpgid`, `syscall.Kill`)
that don't exist on Windows. This was latent: the binary job previously
failed earlier (missing embedded UI), so it never reached `make windows`.

Split the process-group handling into build-tagged files:
- ui/dev_unix.go + ui/dev_windows.go
- cmd/connection_browser_unix.go + cmd/connection_browser_windows.go

Unix keeps the pgid-based group signalling; Windows (which has no POSIX
process groups) signals the process directly. The platform-specific test
assertion is likewise split into _unix_test/_windows_test files.

Verified `go build ./...` for windows, linux, and darwin all pass.
